Added Paul's code and attempted few things.
Would be nice to try and see whether the speeches and when they were said (e.g.: each decade) have a theme that can be extracted or perhaps certain decades resemble a particular theme or topic. I found that almost each cluster has a country associated with it, so perhaps something to look into.